requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
groq
requests
beautifulsoup4
python-dotenv
google-search-results